requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.101.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
mysql-connector-python==8.1.0
mysqlclient==2.2.0
orjson==3.9.10
protobuf==4.21.12
pyasn1==0.5.0
pydantic==2.2.1
pydantic-core==2.6.1
pydantic-settings==2.0.3
PyMySQL==1.1.0
pyodbc==5.0.1
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
SQLAlchemy-serializer==1.4
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.2
